backend/requirements.txt
# PropertyAI Backend Dependencies - Production Ready
# ================================================

# Core FastAPI and Server
fastapi>=0.65.2
uvicorn[standard]>=0.15.0
python-multipart==0.0.9

# HTTP Client
httpx==0.25.2

# JWT Authentication & Security
python-jose[cryptography]==3.3.0

# FastAPI Users
fastapi-users==13.0.0
fastapi-users-db-beanie==3.0.0

# Template Engine
jinja2==3.1.2

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic>=1.10.0,<2.0.0
pydantic-settings>=1.10.0,<2.0.0
email-validator==2.1.0

# MongoDB Support
pymongo==4.6.0
motor==3.3.2
beanie==1.26.0

# File Upload Support
aiofiles==23.2.1
Pillow==10.4.0

# Production Deployment
gunicorn==21.2.0

# Logging & Monitoring
structlog==23.2.0

# Additional Security
cryptography==41.0.7

# Rate Limiting
slowapi==0.1.9
